move history imports and action; changed app layout;
